requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
bleach==2.0.0             # via nbconvert
cycler==0.10.0            # via matplotlib
decorator==4.0.11         # via ipython, traitlets
entrypoints==0.2.2        # via nbconvert
html5lib==0.999999999     # via bleach
ipykernel==4.6.1          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==6.0.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==6.0.0         # via jupyter
jedi==0.10.2              # via ipython
jinja2==2.9.6             # via nbconvert, notebook
jsonschema==2.6.0         # via nbformat
jupyter-client==5.0.1     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.1.0    # via jupyter
jupyter-core==4.3.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
markupsafe==1.0           # via jinja2
matplotlib==2.0.1
mistune==0.7.4            # via nbconvert
nbconvert==5.1.1          # via jupyter, notebook
nbformat==4.3.0           # via ipywidgets, nbconvert, notebook
notebook==5.0.0           # via jupyter, widgetsnbextension
numpy==1.12.1
olefile==0.44             # via pillow
pandocfilters==1.4.1      # via nbconvert
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pillow==4.1.1
prompt-toolkit==1.0.14    # via ipython, jupyter-console
protobuf==3.3.0           # via tensorflow
ptyprocess==0.5.1         # via pexpect, terminado
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.2.0          # via matplotlib
python-dateutil==2.6.0    # via jupyter-client, matplotlib
pytz==2017.2              # via matplotlib
pyzmq==16.0.2             # via jupyter-client
qtconsole==4.3.0          # via jupyter
scikit-learn==0.18.1      # via sklearn
scipy==0.19.0
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via bleach, cycler, html5lib, matplotlib, prompt-toolkit, protobuf, python-dateutil, tensorflow, traitlets
sklearn==0.0
tensorflow==1.1.0
terminado==0.6            # via notebook
testpath==0.3             # via nbconvert
tornado==4.5.1            # via ipykernel, notebook, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
werkzeug==0.12.1          # via tensorflow
wheel==0.29.0             # via tensorflow
widgetsnbextension==2.0.0  # via ipywidgets
